requirements.txt
langchain
langchain-openai
langchain-community
langchainhub
python-dotenv
langchain-ollama
langchain-google-community
streamlit
streamlit-chat 
langchain-deepseek-official
beautifulsoup4
